pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tutorial-agent"
version = "1.0.0"
description = "Interactive multi-agent tutorial system powered by Claude, Tavily, and a local vector store"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "anthropic>=0.40.0",
    "tavily-python>=0.3.3",
    "python-docx>=1.1.2",
    "pdfplumber>=0.11.4",
    "sentence-transformers>=3.1.1",
    "numpy>=1.26",
    "rich>=13.9.4",
    "questionary>=2.0.1",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.1",
]

[project.scripts]
tutorial-agent = "tutorial_agent.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tutorial_agent*"]

[tool.setuptools.package-data]
"*" = [".env.example"]
